ml-pipeline/requirements.txt
# Core ML Libraries
torch==2.1.0
transformers==4.35.0
datasets==2.14.0
tokenizers==0.14.0
accelerate==0.24.0

# Experiment Tracking & Model Management
mlflow==2.8.0
dvc==3.28.0
dvc-s3==2.23.0
optuna==3.4.0
wandb==0.15.12

# Data Processing
pandas==2.1.3
numpy==1.26.0
scikit-learn==1.3.2
scipy==1.11.4
nltk==3.8.1
spacy==3.7.2

# Model Serving
torchserve==0.9.0
torch-model-archiver==0.9.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Monitoring & Evaluation
evidently==0.4.0
prometheus-client==0.18.0
grafana-api==1.0.3
alibi-detect==0.11.4

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
rich==13.6.0
tqdm==4.66.1
joblib==1.3.2

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.0
isort==5.12.0
pre-commit==3.5.0

# Distributed Training (optional)
horovod==0.28.1
ray[train]==2.8.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Vector Operations
faiss-cpu==1.7.4
annoy==1.17.3
hnswlib==0.8.0

# Visualization
matplotlib==3.8.1
seaborn==0.13.0
plotly==5.18.0
tensorboard==2.15.0